Reorganizing code in preparation for RViz merge.
