+++ package.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0"?>
-<package>
+<package format="2">
   <name>or_interactivemarker</name>
   <version>0.0.1</version>
   <description>
@@ -11,14 +11,10 @@
   <author email="mkoval@cs.cmu.edu">Michael Koval</author>
   <license>BSD</license>
   <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>boost</build_depend>
   <build_depend>openrave_catkin</build_depend>
-  <build_depend>interactive_markers</build_depend>
-  <build_depend>std_msgs</build_depend>
-  <build_depend>visualization_msgs</build_depend>
-  <run_depend>boost</run_depend>
-  <run_depend>openrave_catkin</run_depend>
-  <run_depend>interactive_markers</run_depend>
-  <run_depend>std_msgs</run_depend>
-  <run_depend>visualization_msgs</run_depend>
+  <depend>boost</depend>
+  <depend>interactive_markers</depend>
+  <depend>geometry_msgs</depend>
+  <depend>std_msgs</depend>
+  <depend>visualization_msgs</depend>
 </package>
