+++ src/or_interactivemarker.cpp
@@ -3,7 +3,7 @@
 #include <boost/algorithm/string/trim.hpp>
 #include <interactive_markers/interactive_marker_server.h>
 #include "or_interactivemarker.h"
-#include "or_conversions.h"
+#include "util/or_conversions.h"
 
 using boost::format;
 using boost::str;
